add wizard page when submitting job (#3430)
