+++ src/webportal/src/app/job-submission/components/container-size-section.jsx
@@ -58,7 +58,7 @@ export const ContainerSizeSection = (props) => {
   };
 
   return (
-    <BasicSection sectionLabel='Resources per instance' sectionTooltip={PROTOCOL_TOOLTIPS.taskRoleContainerSize}>
+    <BasicSection sectionLabel='Resources' sectionTooltip={PROTOCOL_TOOLTIPS.taskRoleContainerSize}>
       <Stack horizontal gap='l1'>
         <FormShortSection gap='m'>
           <CSpinButton
